+++ consensus/src/test/scala/org/plasmalabs/bridge/stubs/BasePublicApiClientGrpc.scala
@@ -1,0 +1,16 @@
+package org.plasmalabs.bridge.stubs
+
+import cats.effect.IO
+import org.plasmalabs.bridge.consensus.core.PublicApiClientGrpc
+import org.plasmalabs.bridge.consensus.service.StateMachineReply
+import org.plasmalabs.bridge.shared.Empty
+
+class BasePublicApiClientGrpc extends PublicApiClientGrpc[IO] {
+
+  override def replyStartPegin(
+    timestamp:       Long,
+    currentView:     Long,
+    startSessionRes: StateMachineReply.Result
+  ): IO[Empty] = ???
+
+}