+++ consensus/src/test/scala/org/plasmalabs/bridge/stubs/BaseTemplateStorageAlgebra.scala
@@ -1,0 +1,12 @@
+package org.plasmalabs.bridge.stubs
+
+import cats.effect.IO
+import co.topl.brambl.dataApi.{TemplateStorageAlgebra, WalletTemplate}
+
+class BaseTemplateStorageAlgebra extends TemplateStorageAlgebra[IO] {
+
+  override def findTemplates(): IO[Seq[WalletTemplate]] = ???
+
+  override def addTemplate(walletTemplate: WalletTemplate): IO[Int] = ???
+
+}