+++ integration/src/test/scala/xyz/stratalab/bridge/BridgeIntegrationSpec.scala
@@ -1,4 +1,4 @@
-package xyz.stratalab.bridge
+package org.plasmalabs.bridge
 
 import munit.CatsEffectSuite
 
@@ -21,13 +21,15 @@ class BridgeIntegrationSpec
   cleanupDir.test("Bridge should correctly peg-in BTC") { _ =>
     info"Bridge should correctly peg-in BTC" >> successfulPegin()
   }
-  cleanupDir.test("Bridge should fail correctly when user does not send BTC") {
-    _ =>
-      info"Bridge should fail correctly when user does not send BTC" >> failedPeginNoDeposit()
+
+  cleanupDir.test("Bridge should fail correctly when user does not send BTC") { _ =>
+    info"Bridge should fail correctly when user does not send BTC" >> failedPeginNoDeposit()
   }
+
   cleanupDir.test("Bridge should fail correctly when tBTC not minted") { _ =>
     info"Bridge should fail correctly when tBTC not minted" >> failedPeginNoMint()
   }
+
   cleanupDir.test("Bridge should fail correctly when tBTC not redeemed") { _ =>
     info"Bridge should fail correctly when tBTC not redeemed" >> failedRedemption()
   }
@@ -49,7 +51,6 @@ class BridgeIntegrationSpec
   ) { _ =>
     info"Bridge should correctly retry if claim does not succeed" >> successfulPeginWithClaimErrorRetry()
   }
-
 
   cleanupDir.test(
     "Bridge should correctly go back to minting if there is a reorg"