+++ integration/src/test/scala/xyz/stratalab/bridge/FailedRedemptionModule.scala
@@ -1,4 +1,5 @@
-package xyz.stratalab.bridge
+package org.plasmalabs.bridge
+
 import cats.effect.IO
 import org.typelevel.log4cats.syntax._
 
@@ -12,8 +13,8 @@ trait FailedRedemptionModule {
     import cats.implicits._
     assertIO(
       for {
-        newAddress <- getNewAddress
-        _ <- generateToAddress(1, 1, newAddress)
+        newAddress       <- getNewAddress
+        _                <- generateToAddress(1, 1, newAddress)
         txIdAndBTCAmount <- extractGetTxIdAndAmount
         (txId, btcAmount, btcAmountLong) = txIdAndBTCAmount
         startSessionResponse <- startSession(1)
@@ -23,14 +24,14 @@ trait FailedRedemptionModule {
           btcAmount
         )
         signedTxHex <- signTransaction(bitcoinTx)
-        _ <- sendTransaction(signedTxHex)
-        _ <- generateToAddress(1, 8, newAddress)
-        _ <- mintStrataBlock(1, 5)
+        _           <- sendTransaction(signedTxHex)
+        _           <- generateToAddress(1, 8, newAddress)
+        _           <- mintStrataBlock(1, 5)
         _ <-
           (for {
             status <- checkMintingStatus(startSessionResponse.sessionID)
-            _ <- mintStrataBlock(1, 3)
-            _ <- IO.sleep(1.second)
+            _      <- mintStrataBlock(1, 3)
+            _      <- IO.sleep(1.second)
           } yield status)
             .iterateUntil(_.mintingStatus == "PeginSessionStateMintingTBTC")
         _ <- info"We are in the waiting for redemption state"