+++ integration/src/test/scala/xyz/stratalab/bridge/SuccessfulPeginWithClaimReorgModule.scala
@@ -1,4 +1,4 @@
-package xyz.stratalab.bridge
+package org.plasmalabs.bridge
 
 import cats.effect.IO