+++ public-api/src/main/scala/org/plasmalabs/bridge/publicapi/StrataBTCBridgePublicApiParamConfig.scala
@@ -1,0 +1,7 @@
+package org.plasmalabs.bridge.publicapi
+
+import java.io.File
+
+case class StrataBTCBridgePublicApiParamConfig(
+  configurationFile: File = new File("application.conf")
+)