+++ public-api/src/main/scala/xyz/stratalab/bridge/publicapi/StrataBTCBridgePublicApiParamConfig.scala
@@ -1,7 +1,0 @@
-package xyz.stratalab.bridge.publicapi
-
-import java.io.File
-
-case class StrataBTCBridgePublicApiParamConfig(
-  configurationFile: File = new File("application.conf")
-)